brainTumorVA/requirements.txt
ultralytics>=8.0.0
langchain>=0.1.0
langchain-groq
langchain-community
langchain-core
faiss-cpu
python-dotenv
pillow
fpdf2
opencv-python
numpy
matplotlib
streamlit
torch